#127 refactor DdrWorkflowBase test
