+++ src/main/kotlin/org/objectionary/ddr/sources/Sources.kt
@@ -37,21 +37,21 @@ class SrsTransformed(
 ) : Sources {
     private val logger = LoggerFactory.getLogger(this.javaClass.name)
 
-    /** @property path path to directory with source files or single source file */
-    val path: Path = Path.of(strPath)
+    /** @property inPath path to directory with source files or single source file */
+    val inPath: Path = Path.of(strPath)
 
     /** @property documents all documents */
     private val documents: MutableMap<Document, String> = mutableMapOf()
 
     /**
-     * Walks through [path], make some xsl transformation on xmir files using [transformer] and collect [documents]
+     * Walks through [inPath], make some xsl transformation on xmir files using [transformer] and collect [documents]
      *
      * @return [MutableMap] with collected [documents] as key and their filenames as value
      *
      * @todo #121:30min `createTempDirectories` method is redundantly called in a loop. This method should be called once.
      */
     override fun walk(): MutableMap<Document, String> {
-        Files.walk(path)
+        Files.walk(inPath)
             .filter(Files::isRegularFile)
             .forEach {
                 val tmpPath = createTempDirectories(it.toString())
@@ -107,7 +107,7 @@ class SrsTransformed(
      *
      */
     private fun generateTmpPath(filename: String): String {
-        val strPath = path.toString()
+        val strPath = inPath.toString()
         return "${strPath.substringBeforeLast(sep)}$sep${strPath.substringAfterLast(sep)}_$postfix${filename.substring(strPath.length)}"
     }
 }
