+++ src/test/kotlin/org/objectionary/ddr/integration/IntegrationTestBase.kt
@@ -11,4 +11,11 @@ interface IntegrationTestBase : TestBase {
 
     override fun constructOutPath(directoryName: String): String =
         "src${sep}test${sep}resources${sep}integration${sep}out$sep$directoryName"
+
+    /**
+     * @param directoryName name of the result directory
+     * @return path to output location
+     */
+    fun constructResultPath(directoryName: String) =
+        "${constructInPath(directoryName).substringBeforeLast(sep)}$sep${directoryName}_$postfix"
 }
