+++ src/test/kotlin/org/objectionary/ddr/unit/graph/builder/BuilderBase.kt
@@ -48,7 +48,7 @@ open class BuilderBase : UnitTestBase {
         val expected = File(constructOutPath(testName)).bufferedReader().readText()
         logger.debug(actual)
         checkOutput(expected, actual)
-        deleteTempDir(sources.path)
+        deleteTempDir(sources.inPath)
     }
 
     override fun constructOutPath(directoryName: String): String =
